Move timegm.h to include/exiv2 since it is used by the app
