+++ exiv2lib/include/exiv2/timegm.h
